Separate implementations in pkg
